% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adeliepenguin}
\alias{adeliepenguin}
\title{Adelie penguin list of matrices}
\format{
Object of class \code{list} of \code{matrices}
}
\usage{
data("adeliepenguin")
}
\description{
List of Matrix Projection Models for the Adelie penguin \emph{pygoscelis adeliae}
based on 2 age classes. The matrices are based on a population
situated in the Copacabana colony on King George Island, Antarctica.
The matrices are based on two stage classes: \cr
Age 1: Juveniles  (ages 0–2)\cr
Age 2: Adults  (ages 3+)\cr
}
\examples{
  # read in data and view
  data(adeliepenguin)
  adeliepenguin

}
\references{
\itemize{
\item  Hinke, J. T., S. G. Trivelpiece, and W. Z. Trivelpiece. 2017.
Variable vital rates and the risk of population declines in Adélie penguins
from the Antarctic Peninsula region.
Ecosphere 8(1):e01666. doi: 10.1002/ecs2.1666
}
}
\concept{Adelie}
\concept{data}
\concept{penguin}
\concept{pygoscelis adeliae}
\keyword{datasets}
