% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blue_crane}
\alias{blue_crane}
\title{Blue crane list of matrices}
\format{
Object of class \code{list} of \code{matrices}
}
\usage{
data("blue_crane")
}
\description{
List of Matrix Projection Models for the blue crane \emph{Gopherus agassizii}
based on 5 age classes. The matrices are based on a population centred on
South Africa's grassland, eastern Nama Karoo and Overberg regions.
The matrices are based on age:
Age 1: Year 1\cr
Age 2: Year 2\cr
Age 4: Year 4\cr
Age 5: Year 5\cr
}
\examples{
  # read in data and view
  data(blue_crane)
  blue_crane

}
\references{
\itemize{
\item Altwegg, R. and Anderson, M.D. (2009),
Rainfall in arid zones: possible effects of climate change on the population
ecology of blue cranes. Functional Ecology, 23: 1014-1021.
https://doi.org/10.1111/j.1365-2435.2009.01563.x.
}
}
\concept{Anthropoides paradiseus}
\concept{blue}
\concept{crane}
\concept{data}
\keyword{datasets}
