% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_dist_function.R
\name{demres_dist}
\alias{demres_dist}
\title{Provides time-varying and time-constant resilience metrics for animal
populations}
\usage{
demres_dist(table, metric, measure)
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated####CHANGE ALL OF IT HERE
with the demres function}

\item{metric:}{"reac": Reactivity: first-timestep amplification
and first-timestep attenuation for a population matrix
projection model.
"inertia": Population inertia for a population
matrix projection model.
"dr": Damping ratio of a given population
matrix projection model.
"maxamp": Maximal amplification for a population
matrix projection model.
"maxatt": Maximal attenuation for a population
matrix projection model.}

\item{measure:}{"RMSE"
"rRMSE"
"MAPE"}
}
\value{
A dataframe containing all the resilience metrics
}
\description{
\code{demres_dist} calculates resilience metrics of a population based
on a list of matrix population models
}
\details{
This function applies the function "calc_resilience" to a list of matrices
and returns either time-varying metrics or time-constant metrics
}
\examples{
\dontrun{

#load packages 
library(Rcompadre)
library(dplyr)
library(popdemo)

# load data 
comadre <- cdb_fetch("comadre")

#selecting the blue crane 
blue_crane <- comadre \%>\% dplyr::filter(SpeciesAccepted  == "Anthropoides paradiseus")

#extracting matrices
blue_crane_matA <- Rcompadre::matA(blue_crane)

# simulate an initial vector
Cranevec1 <- runif(5)
Cranevec1 <- Cranevec1/sum(Cranevec1) #scales the vec to sum to 1


BC_TVTC_demres <-
  demres(
    blue_crane_matA,
    metrics = "all",
    bounds = TRUE,
    initvec = Cranevec1,
    popname = "blue crane",
    time = "both"
  )
  
dist_BC <- demres_dist(table = BC_TVTC_demres, metric = "inertia", measure = "all")

}
}
