% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TD_initvec_function.R
\name{get_TD_initvec}
\alias{get_TD_initvec}
\title{Helper function for setting up a list ot´f time dependant initial vectors,
starting with one initial vector}
\usage{
get_TD_initvec(initvec, listA)
}
\arguments{
\item{initvec}{a one-column matrix describing the age/stage distribution
('demographic structure') used to calculate a 'case-specific',
stage age structure for the first year. The other initial vectors will be
calculated from this first initial vector.}

\item{listA}{a list of square, primitive, irreducible, non-negative numeric
matrices of any dimension}
}
\description{
\code{get_TD_initvec} provides a list of initial vectors that are projected for
each year with each time-varying Matrix, starting with one initial vector
}
\keyword{internal}
