% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_function.R
\name{plot.resil}
\alias{plot.resil}
\title{Plotting method for objects of class resil}
\usage{
\method{plot}{resil}(x, rRMSE = FALSE, RMSE = FALSE, MAPE = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
the argument \code{x} (invisibly)
}
\description{
Plotting method for objects of class resil
}
