% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testplot.R
\name{plot_test}
\alias{plot_test}
\title{Plot test}
\usage{
plot_test(ask = interactive())
}
\arguments{
\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics:par]{par(ask=.)}}.}
}
\description{
Plot test
}
