% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_function.R
\name{print.resil}
\alias{print.resil}
\title{Printing method for objects of class resil}
\usage{
\method{print}{resil}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
the argument \code{x} (invisibly)
}
\description{
Printing method for objects of class resil
}
