% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{time_to_target}
\alias{time_to_target}
\title{Calculate Time to Target}
\usage{
time_to_target(A, n0, target, max_time = 10000, chunk = 100)
}
\arguments{
\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}
}
\description{
Internal functions used by \code{\link[=calc_resilience]{calc_resilience()}}.
}
\seealso{
\code{\link[=calc_resilience]{calc_resilience()}}
}
\keyword{internal}
