% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{calc_convt}
\alias{calc_convt}
\title{Calculate convergence time}
\usage{
calc_convt(metrics, vector, A, bounds, accuracy, iterations)
}
\arguments{
\item{metrics}{"reac": Calculates reactivity: first-timestep amplification
and first-timestep attenuation for a matrix
projection model. \cr
"inertia": Calculates population inertia for a
matrix projection model.\cr
"dr": Calculates the damping ratio of a given
matrix projection model.\cr
"maxamp": Calculates maximal amplification for a
matrix projection model.\cr
"maxatt": Calculates maximal attenuation for a
matrix projection model.\cr
"convt": Calculates the time to convergence of a
matrix projection model.\cr
"all": all of the above metrics are provided.}

\item{vector}{a numeric vector or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
resilience metric}

\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{bounds}{(optional) Boolean. Set to FALSE as default. If TRUE, specifies whether the upper and  lower
bound should be calculated. If initial vector is not specified, the function
provides metrics at their upper and lower bounds, calculated based on the stage-biased vector.
If vector is specified, the function provides also the metrics calculated
based on the initial vector.}

\item{accuracy}{the accuracy with which to determine convergence to asymptotic growth,
expressed as a proportion. Set to 0.01 by default.}

\item{iterations}{the maximum number of iterations of the model before the code breaks. For slowly-converging models
and/or high specified convergence accuracy, this may need to be increased.
Set to 1e+05 by default.
\strong{Vik}: the option above (iterations) is used for convergence time only, right? Should be mentioned then}
}
\description{
Internal functions used by \code{\link[=calc_resilience]{calc_resilience()}}.
}
\seealso{
\code{\link[=calc_resilience]{calc_resilience()}}
}
