% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{calc_maxamp_or_maxatt}
\alias{calc_maxamp_or_maxatt}
\title{Calculate reactivity or inertia metric}
\usage{
calc_maxamp_or_maxatt(
  metrics,
  vector,
  A,
  return.N = return.N,
  return.t = return.t
)
}
\arguments{
\item{metrics}{"convt": Calculates the time to convergence of a
matrix projection model.\cr
"dr": Calculates the damping ratio of a given
matrix projection model.\cr
"maxamp": Calculates maximal amplification for a
matrix projection model.\cr
"maxatt": Calculates maximal attenuation for a
matrix projection model.\cr
"reac": Calculates reactivity: first time step amplification
and first time step attenuation for a matrix
projection model.\cr
"all": all of the above metrics are provided.}

\item{vector}{a numeric vector or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
resilience metric}

\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{return.N}{Boolean. Set to TRUE as default. If TRUE returns population size at the point of the metric reached,
if set to FALSE, returns the standardised maximal amplification}

\item{return.t}{Boolean. If TRUE, returns the time at which the meric occurs in the population projection}
}
\description{
Internal functions used by \code{\link[=calc_resilience]{calc_resilience()}}.
}
\seealso{
\code{\link[=calc_resilience]{calc_resilience()}}

Calculate maxam or maxatt metric

Internal functions used by \code{\link[=calc_resilience]{calc_resilience()}}.

\code{\link[=calc_resilience]{calc_resilience()}}
}
