% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{calc_maxamp_or_maxatt}
\alias{calc_maxamp_or_maxatt}
\title{Calculate maxam or maxatt metric}
\usage{
calc_maxamp_or_maxatt(metrics, vector, A, bounds)
}
\arguments{
\item{metrics}{"reac": Calculates reactivity: first-timestep amplification
and first-timestep attenuation for a population matrix
projection model.
"inertia": Calculates population inertia for a population
matrix projection model.
"dr": Calculate the damping ratio of a given population
matrix projection model.
"maxamp": Calculate maximal amplification for a population
matrix projection model.
"maxatt": Calculate maximal attenuation for a population
matrix projection model.
"all": all of the above metrics are provided}

\item{vector}{a numeric vector or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
maximal amplification}

\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{bounds}{(optional) if TRUE, specifies whether the upper and  lower bound
should be calculated
if vector is not specified, the function provides metrics in their upper and
lower bound, calculated based on the stage-biased vector
if vector is specified, the function provides also the metrics calculated based
on the inital vector}
}
\description{
Internal functions used by \code{\link[=calc_resilience]{calc_resilience()}}.
}
\seealso{
\code{\link[=calc_resilience]{calc_resilience()}}
}
