% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{calc_reac_or_inertia}
\alias{calc_reac_or_inertia}
\title{Calculate reactivity or inertia metric}
\usage{
calc_reac_or_inertia(metrics, vector, A, bounds)
}
\arguments{
\item{metrics}{"reac": Calculates reactivity: first-timestep amplification
and first-timestep attenuation for a matrix
projection model. \cr
"inertia": Calculates population inertia for a
matrix projection model.\cr
"dr": Calculates the damping ratio of a given
matrix projection model.\cr
"maxamp": Calculates maximal amplification for a
matrix projection model.\cr
"maxatt": Calculates maximal attenuation for a
matrix projection model.\cr
"convt": Calculates the time to convergence of a
matrix projection model.\cr
"all": all of the above metrics are provided.}

\item{vector}{a numeric vector or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
resilience metric}

\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{bounds}{(optional) Boolean. Set to FALSE as default. If TRUE, specifies whether the upper and  lower
bound should be calculated. If initial vector is not specified, the function
provides metrics at their upper and lower bounds, calculated based on the stage-biased vector.
If vector is specified, the function provides also the metrics calculated
based on the initial vector.}
}
\description{
Internal functions used by \code{\link[=calc_resilience]{calc_resilience()}}.
}
\seealso{
\code{\link[=calc_resilience]{calc_resilience()}}
}
