% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{calc_resilience}
\alias{calc_resilience}
\title{Provides resilience metrics from the package \emph{{popdemo}} for one matrix}
\usage{
calc_resilience(
  A,
  metrics = "all",
  vector = vect,
  popname = NULL,
  verbose = TRUE,
  accuracy = 0.01,
  iterations = 1e+05,
  return.N = TRUE,
  return.t = TRUE
)
}
\arguments{
\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{metrics}{"convt": Calculates the time to convergence of a
matrix projection model.\cr
"dr": Calculates the damping ratio of a given
matrix projection model.\cr
"inertia": Calculates population inertia for a
matrix projection model.\cr
"maxamp": Calculates maximal amplification for a
matrix projection model.\cr
"maxatt": Calculates maximal attenuation for a
matrix projection model.\cr
"reac": Calculates reactivity: first time step amplification
and first time step attenuation for a matrix
projection model.\cr
"all": all of the above metrics are provided.}

\item{vector}{a numeric vector or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
resilience metric}

\item{popname}{a character string describing the name of the population.}

\item{verbose}{Boolean. Set to TRUE as default. Indicates whether the messages about failure
to compute particular metric should be displayed or not (default = TRUE)}

\item{accuracy}{option for calculating convergence time: the accuracy with which to determine convergence to asymptotic growth,
expressed as a proportion. Set to 0.01 by default.}

\item{iterations}{option for calculating convergence time: the maximum number of iterations of the model. Set to 1e+05 by default. For slowly-converging models
and/or high specified convergence accuracy, this may need to be increased.}

\item{bounds}{(optional) Boolean. Set to FALSE as default. If TRUE, specifies whether the upper and  lower
bound should be calculated. If initial vector is not specified, the function
provides metrics at their upper and lower bounds, calculated based on the stage-biased vector.
If vector is specified, the function provides also the metrics calculated
based on the initial vector.}
}
\value{
A vector containing all the resilience metrics
}
\description{
\code{calc_resilience} calculates resilience metrics of a population based
on a matrix population model
}
\details{
This function computes different metrics of resilience, given by the \emph{{popdemo}}
package based on a matrix projection model.
}
\examples{
data(adeliepenguin)

set.seed(1234)
penguinvec1 <- c(25, 75)
penguin1 <- adeliepenguin[[1]]

all_penguin_demres <- calc_resilience(penguin1, metrics = c("all"),
vector = penguinvec1, popname = "adelie penguin", verbose = TRUE, return.N = TRUE, return.t = TRUE)

}
\keyword{internal}
