% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resilience_function.R
\name{calc_resilience}
\alias{calc_resilience}
\title{Provides resilience metrics from the package popdemo}
\usage{
calc_resilience(A, metrics, bounds = FALSE, initvec = "n", popname = NULL)
}
\arguments{
\item{A}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{bounds}{(optional) if TRUE, specifies whether the upper and  lower bound
should be calculated
if vec is not specified, the function provides metrics in their upper and
lower bound, calculated based on the stage-biased vec
if vec is specified, the function provides also the metrics calculated based
on the inital vec}

\item{initvec}{a numeric vec or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
maximal amplification}

\item{popname}{a character string describing the name of the population}

\item{metrics:}{"reac": Calculates reactivity: first-timestep amplification
and first-timestep attenuation for a population matrix
projection model.
"inertia": Calculates population inertia for a population
matrix projection model.
"dr": Calculate the damping ratio of a given population
matrix projection model.
"maxamp": Calculate maximal amplification for a population
matrix projection model.
"maxatt": Calculate maximal attenuation for a population
matrix projection model.
"all": all of the above metrics are provided}
}
\value{
A vector containing all the resilience metrics
}
\description{
\code{calc_resilience} calculates resilience metrics of a population based
on a matrix population model
}
\details{
This function compiles different metrics of resilience, given by the popdemo
package based on a population matrix projection model.
}
\examples{
\dontrun{
library(popdemo)
data(Tort)

Tortvec1 <- runif(8) # create initial vec
Tortvec1 <- Tortvec1/sum(Tortvec1) #scales the vec to sum to 1

all_tort_demres <- calc_resilience(Tort, metrics = c("all"), 
initvec = Tortvec1, bounds = TRUE, popname = "Tortoise")

}
}
