% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_function.R
\name{demres}
\alias{demres}
\title{Provides resilience metrics from the package popdemo}
\arguments{
\item{listA}{a square, primitive, irreducible, non-negative numeric matrix of any
dimension}

\item{metrics:}{"reac": Calculates reactivity: first-timestep amplification
and first-timestep attenuation for a population matrix
projection model.
"inertia": Calculates population inertia for a population
matrix projection model.
"dr": Calculate the damping ratio of a given population
matrix projection model.
"maxamp": Calculate maximal amplification for a population
matrix projection model.
"maxatt": Calculate maximal attenuation for a population
matrix projection model.
"all": all of the above metrics are provided}

\item{bounds}{(optional) if TRUE, specifies whether the upper and  lower bound
should be calculated
if vec is not specified, the function provides metrics in their upper and
lower bound, calculated based on the stage-biased vec
if vec is specified, the function provides also the metrics calculated based
on the inital vec}

\item{initvec}{a numeric vec or one-column matrix describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific'
maximal amplification}

\item{popname}{a character string describing the name of the population}

\item{time}{a character string: "constant" or "varying"
"constant" : if the metrics are to be calculated over the whole study period
"varying": if the metrics are to be calculated for each time step}
}
\value{
A tibble containing all the resilience metrics
}
\description{
\code{demres} calculates resilience metrics of a population based
on a list of matrix population models
}
\details{
This function first checks that the input is a list of matrices,
if not, it gives back an error message a
if yes it applies the function "wrapper_res_met" to a list of matrices
}
