% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_plot_function.R
\name{demres_plot}
\alias{demres_plot}
\title{Plot the resulting demographic resilience metrics}
\usage{
demres_plot(
  metric,
  table,
  plotname = paste0(getwd(), "/plot_demres_", metric, ".pdf"),
  rRMSE = FALSE,
  RMSE = FALSE,
  MAPE = FALSE
)
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated
with the demres function}

\item{metric:}{"reac": Reactivity: first-timestep amplification
and first-timestep attenuation for a population matrix
projection model.
"inertia": Population inertia for a population
matrix projection model.
"dr": Damping ratio of a given population
matrix projection model.
"maxamp": Maximal amplification for a population
matrix projection model.
"maxatt": Maximal attenuation for a population
matrix projection model.}

\item{RMSE:}{(optional) if TRUE: calculates the RMSE (sqrt(mean((TV-TC)^2))
with TV: the time-Varying resilience metric and TC the time constant one)
and adds it on the plot.}

\item{rRMSE:}{(optional) if TRUE: calculates the relative RMSE
(sqrt(mean((TV-TC)^2)) / sd(TV) with TV: the time-Varying resilience metric
and TC the time constant ) and adds it on the plot}

\item{MAPE:}{(optional) if TRUE: calculates the MAPE (mean(abs(TV - TC)))
and adds it on the plot}

\item{plotname:}{name of the plot with directory. By default:
plotname = paste0(getwd(), "/plot_demres_", metric, ".pdf")}
}
\value{
A plot displaying the chosen metric along a time axis
}
\description{
\code{demres_plot} provides a plot to visually inspect the resilience metric
along a time axis
}
