% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_plot_function.R
\name{demres_plot}
\alias{demres_plot}
\title{Plot the resulting demographic resilience metrics}
\usage{
demres_plot(table, ask = interactive())
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated
with the \code{resilience} function}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics:par]{par(ask=.)}}}
}
\value{
A plot displaying the chosen metric(s) along a time axis
}
\description{
The function \code{plot} calls \code{demres_plot} to visually inspect the resilience metric(s)
along a time axis
}
\examples{
# load data
data(adeliepenguin)

# simulate an initial vector
set.seed(1234)
penguinvec1 <- runif(5)
penguinvec1 <- penguinvec1/sum(penguinvec1) #scales the vec to sum to 1


AP_TVTC_demres <-
  resilience(
    listA = adeliepenguin,
    metrics = "all",
    bounds = TRUE,
    vector = penguinvec1,
    TDvector = FALSE,
    popname = "adelie penguin",
    time = "both",
    verbose = TRUE
  )

plot(AP_TVTC_demres)
}
