% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_plot_function.R
\name{demres_plot}
\alias{demres_plot}
\title{Plot the resulting demographic resilience metrics}
\usage{
demres_plot(table, ask = interactive())
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated
with the \code{resilience} function}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics:par]{par(ask=.)}}}
}
\value{
A plot displaying the chosen metric(s) along a time axis
}
\description{
The function \code{plot} calls \code{demres_plot} to visually inspect the resilience metric(s)
along a time axis
}
\examples{
# load data
data(bluecrane)

# simulate an initial vector
set.seed(1234)
Cranevec1 <- runif(5)
Cranevec1 <- Cranevec1/sum(Cranevec1) #scales the vec to sum to 1


BC_TVTC_demres <-
  resilience(
    listA = bluecrane,
    metrics = "all",
    bounds = TRUE,
    vector = Cranevec1,
    TDvector = FALSE,
    popname = "blue crane",
    time = "both",
    verbose = TRUE
  )

plot(BC_TVTC_demres)
}
