% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_plot_function.R
\name{demres_plot}
\alias{demres_plot}
\title{Plot the resulting demographic resilience metrics}
\usage{
demres_plot(table)
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated
with the resilience function}
}
\value{
A plot displaying the chosen metric along a time axis
}
\description{
\code{demres_plot} provides a plot to visually inspect the resilience metric
along a time axis
}
\examples{
# load data
data(bluecrane)

# simulate an initial vector
set.seed(1234)
Cranevec1 <- runif(5)
Cranevec1 <- Cranevec1/sum(Cranevec1) #scales the vec to sum to 1


BC_TVTC_demres <-
  resilience(
    bluecrane,
    metrics = "all",
    bounds = TRUE,
    vector = Cranevec1,
    popname = "blue crane",
    time = "both"
  )

plot(BC_TVTC_demres)
}
