% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_summary_function.R
\name{demres_summary}
\alias{demres_summary}
\title{Calculates the distance between resilience metrics calculated with time-varying
and time-constant approaches}
\usage{
demres_summary(table, fn = list(mean, sd))
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated
with the resilience function}
}
\value{
A data frame displaying the distance measures for the metrics that are present in the table
}
\description{
The function \code{summary} calls \code{demres_summary} to calculate distance between resilience metrics calculated with
time-varying and time-constant approaches:
"RMSE": sqrt(mean((TV-TC)^2))
with TV: the time-varying resilience metric and TC the time constant one
"rRMSE": sqrt(mean((TV-TC)^2)) / sd(TV)
with TV: the time-varying resilience metric and TC the time constant
"MAPE": mean(abs(TV - TC))/TC
with TV: the resilience metric calculated using the time-varying approach
and TC - using the time-constant approach.
}
\examples{
# load data
data(adeliepenguin)

# simulate an initial vector
set.seed(125435)
penguinvec_list_rel <- replicate(n =28, expr= runif(2))
penguinvec_list_norm <- lapply(1:28, FUN = function(x){penguinvec_list_rel[, x]/ sum(penguinvec_list_rel[, x])})
penguinvec_list <- lapply(penguinvec_list_norm, FUN = function(x){round(x*100)})

AP_demres <-
  resilience(
    listA = adeliepenguin,
    metrics = "all",
    vector = penguinvec_list,
    TDvector = FALSE,
    popname = "adelie penguin",
    verbose = TRUE,
    return.N = TRUE,
    return.t = TRUE
  )


summary_AP <- summary(AP_TVTC_demres, fn = list(mean, sd, var))

It is also possible to compile your own functions, for example

coeffvar <- function(data, na.rm = TRUE){ # have to soecify na.rm = TRUE here
CV <- sd(data, na.rm = TRUE) / mean(data, na.rm = TRUE) * 100
}

AP_CV <- summary(AP_TVTC_demres, fn = list(coeffvar))

Or a combination
AP_mix <- summary(AP_TVTC_demres, fn = list(coeffvar, mean, sd))


}
