% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_summary_function.R
\name{demres_summary}
\alias{demres_summary}
\title{Calculates the requested summary statistics for time-varying resilience metrics}
\usage{
demres_summary(table, fn = list(mean, sd))
}
\arguments{
\item{table}{A dataframe containing all the resilience metrics calculated
with the resilience function}
}
\value{
A data frame with requested summary statistics (in columns) per each
resilience metric.
}
\description{
The function \code{summary} calls \code{demres_summary} to calculate specified summary
statistics for time-varying resilience metrics. By default mean and sd are
calculated.
}
\examples{
# load data
data(adeliepenguin)

# simulate an initial vector
set.seed(125435)
vec <- round(runif(2) * 100 )
vector_list <- lapply(1:28, function(i) vec)

AP_demres <-
  resilience(
    listA = adeliepenguin,
    metrics = "all",
    vector = vector_list,
    TDvector = FALSE,
    popname = "adelie penguin",
    verbose = TRUE,
    return.N = TRUE,
    return.t = TRUE
  )


summary_AP <- summary(AP_demres, fn = list(mean, sd, var))

It is also possible to compute your own functions, for example

coeffvar <- function(data, na.rm = TRUE){ # have to specify na.rm = TRUE here
CV <- sd(data, na.rm = TRUE) / mean(data, na.rm = TRUE) * 100
}

AP_CV <- summary(AP_demres, fn = list(coeffvar))

Or a combination
AP_mix <- summary(AP_demres, fn = list(coeffvar, mean, sd))


}
