% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TD_initvec_function.R
\name{get_TD_vector}
\alias{get_TD_vector}
\title{Helper function for setting up a list of time-dependent initial vectors,
starting with the first initial vector}
\usage{
get_TD_vector(IV, listA)
}
\arguments{
\item{IV}{a one-column matrix describing the age/stage distribution
('demographic structure') used to calculate a 'case-specific',
stage/ age structure for the first year. The other initial vectors will be
calculated from this first initial vector.}

\item{listA}{a list of square, primitive, irreducible, non-negative numeric
matrices of any dimension}
}
\description{
\code{get_TD_vector} provides a list of initial vectors that are used for each time
step together with each time-varying population matrix to project population
dynamics over time.
}
\keyword{internal}
