% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_plot_function.R
\name{help_plot}
\alias{help_plot}
\title{Helper function for the demres_plot function}
\usage{
help_plot(metric, table)
}
\arguments{
\item{metric}{"reac": Calculates reactivity: first-timestep amplification
and first-timestep attenuation for a matrix
projection model. \cr
"inertia": Calculates population inertia for a
matrix projection model.\cr
"dr": Calculates the damping ratio of a given
matrix projection model.\cr
"maxamp": Calculates maximal amplification for a
matrix projection model.\cr
"maxatt": Calculates maximal attenuation for a
matrix projection model.\cr
"convt": Calculates the time to convergence of a
matrix projection model.\cr
"all": all of the above metrics are provided.}

\item{table}{A dataframe containing all the resilience metrics calculated
with the resilience function}
}
\value{
A plot displaying the chosen metric along a time axis
}
\description{
\code{help_plot} provides a plot to visually inspect the resilience metric
along a time axis
}
\keyword{internal}
