% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_functions}
\alias{helper_functions}
\alias{RMSE}
\title{Helper functions for demres_dist function from the package "demres"}
\usage{
RMSE(TV, TC)
}
\arguments{
\item{TV}{A vector containing values of time-varying demographic resilience
for one specific  metric, calculated with demres}

\item{TC}{A vector containing the value of the time-constant demographic
resilience for the same one specific  metric as for TV, calculated with demres}
}
\value{
Values of the distance measure
}
\description{
\code{helper_functions} Calculates distance between the time-varying resilience
metric and the time-constant one
}
