% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_function.R
\name{plot.resil}
\alias{plot.resil}
\title{Plotting method for objects of class resil}
\usage{
\method{plot}{resil}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
plots
}
\description{
Plotting method for objects of class resil
}
