% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_function.R
\name{plot.resil}
\alias{plot.resil}
\title{Plotting method for objects of class resil}
\usage{
\method{plot}{resil}(x, rRMSE = FALSE, RMSE = FALSE, MAPE = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{rRMSE}{(optional) if TRUE: calculates the relative RMSE
(sqrt(mean((TV-TC)^2)) / sd(TV) with TV: the time-Varying resilience metric
and TC the time constant ) and adds it on the plot}

\item{RMSE}{(optional) if TRUE: calculates the RMSE (sqrt(mean((TV-TC)^2))
with TV: the time-Varying resilience metric and TC the time constant one)
and adds it on the plot.}

\item{MAPE}{(optional) if TRUE: calculates the MAPE ((mean(abs(TV - TC))/100)
and adds it on the plot}

\item{...}{further arguments passed to or from other methods}
}
\value{
the argument \code{x} (invisibly)
}
\description{
Plotting method for objects of class resil
}
