% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_general.R
\name{plot_general}
\alias{plot_general}
\title{Helper function for the demres_plot function}
\usage{
plot_general(metric, table)
}
\arguments{
\item{metric}{"convt": the time to convergence of a
matrix projection model.\cr
"dr": the damping ratio of a given
matrix projection model.\cr
"inertia": population inertia for a
matrix projection model.\cr
"maxamp": maximal amplification for a
matrix projection model.\cr
"maxatt": maximal attenuation for a
matrix projection model.\cr
"reac": reactivity: first time step amplification
and first time step attenuation for a matrix
projection model.\cr}

\item{table}{A dataframe containing all the resilience metrics calculated
with the resilience function}
}
\value{
A plot displaying the chosen metric along a time axis
}
\description{
\code{plot_general} provides a plot to visually inspect the resilience metric
along a time axis
}
\keyword{internal}
