% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_function.R
\name{resilience}
\alias{resilience}
\title{Calculates time-varying resilience metrics for
populations based on a list of matrix population models}
\usage{
resilience(
  listA,
  metrics = "all",
  vector,
  TDvector = FALSE,
  popname = NULL,
  verbose = TRUE,
  accuracy = 0.01,
  iterations = 1e+05,
  return.N = TRUE,
  return.t = TRUE
)
}
\arguments{
\item{listA}{a list of square, primitive, irreducible, non-negative numeric
matrices of any dimension}

\item{metrics}{"convt": Calculates the time to convergence of a
matrix projection model.\cr
"dr": Calculates the damping ratio of a given
matrix projection model.\cr
"maxamp": Calculates maximal amplification for a
matrix projection model.\cr
"maxatt": Calculates maximal attenuation for a
matrix projection model.\cr
"reac": Calculates reactivity: first time step amplification
and first time step attenuation for a matrix
projection model.\cr
"all": all of the above metrics are provided.}

\item{vector}{a list of numeric vectors or one-column matrices describing the age/stage
distribution ('demographic structure') used to calculate a 'case-specific' resilience metric,
based on the stage- or age-structure.}

\item{TDvector}{Boolean. Set to FALSE as default. Specifies whether or not the
user wants to use, instead of specified population vectors, so-called "time-dependent" vectors that
are obtained from the projection of the population to the current time step using
the specified matrix for each time step.}

\item{popname}{a character string describing the name of the population.}

\item{verbose}{Boolean. Set to TRUE as default. Indicates whether the messages about failure
to compute particular metric should be displayed or not (default = TRUE)}

\item{accuracy}{option for calculating convergence time: the accuracy with which to determine convergence to asymptotic growth,
expressed as a proportion. Set to 0.01 by default.}

\item{iterations}{option for calculating convergence time: the maximum number of iterations of the model. Set to 1e+05 by default. For slowly-converging models
and/or high specified convergence accuracy, this may need to be increased.}

\item{return.N}{Boolean. Set to TRUE as default. If TRUE returns population size.
If set to FALSE, returns the standardised value of the requested metric.}

\item{return.t}{Boolean. If TRUE, returns the time at which the metric is reached in the population projection
If FALSE does not return the time.}
}
\value{
An object of class "resil", which is a dataframe
containing the requested resilience metrics.
}
\description{
This function works with a list of matrices or just with one matrix
and returns time-varying metrics.
}
\examples{
# load data
data(adeliepenguin)

# simulate an initial vector
set.seed(125435)
vec <- round(runif(2) * 100 )
vector_list <- lapply(1:28, function(i) vec)

AP_demres <-
  resilience(
    listA = adeliepenguin,
    metrics = "all",
    vector = vector_list,
    TDvector = FALSE,
    popname = "adelie penguin",
    verbose = TRUE,
    return.N = TRUE,
    return.t = TRUE
  )

AP_demres1 <-
  resilience(
    listA = list(adeliepenguin[[1]]),
    metrics = "all",
    vector = list(vector_list[[1]]),
    TDvector = FALSE,
    popname = "adelie penguin",
    verbose = TRUE,
    return.N = TRUE,
    return.t = TRUE
  )

# an example with time-dependent population vector
AP_demres_TD <-
  resilience(
    listA = adeliepenguin,
    metrics = "all",
    vector = vector_list,
    TDvector = TRUE,
    popname = "adelie penguin",
    verbose = TRUE,
    return.N = TRUE,
    return.t = TRUE
  )

}
