% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_function.R
\name{summary.resil}
\alias{summary.resil}
\title{Summary method for objects of class resil}
\usage{
\method{summary}{resil}(object, f = "wide", ...)
}
\arguments{
\item{object}{an object used to select a method}

\item{f}{A character specifying whether the output should be shown in
"long" (demographic resilience metrics as row names) or in "wide" (demographic
resilience metrics as column names) format. Defaults to "wide".}

\item{...}{further arguments passed to or from other methods}
}
\value{
summary statistics
}
\description{
Summary method for objects of class resil
}
\seealso{
\code{\link[=demres_dist]{demres_dist()}} for details
}
