% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_function.R
\name{summary.resil}
\alias{summary.resil}
\title{Summary method for objects of class resil}
\usage{
\method{summary}{resil}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
summary statistics
}
\description{
Summary method for objects of class resil
}
