% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demres_summary_function.R
\name{summary.temp}
\alias{summary.temp}
\title{Calculate maxam or maxatt metric}
\usage{
\method{summary}{temp}(x, fn = list(mean, sd))
}
\description{
Internal functions used by \code{\link[=demres_summary]{demres_summary()}}.
}
\seealso{
\code{\link[=demres_summary]{demres_summary()}}
}
